### 2024-02-07 #6335 @Kevin-Mattheus-Moerman
@b-pardi thanks for this submission. As AEiC I will now help with the initial steps for this paper.

The above paper compilation error relates to the red highlighted "tabs"/indentation issues in your paper YAML:

![image](https://github.com/openjournals/joss-reviews/assets/8392709/019b46ff-71bf-4745-af8e-a5c481ce14c7)

Can you fix these? You can call `@editorialbot generate pdf` to try the compilation of the draft PDF again.
