### 2024-02-20 #6335 @b-pardi
Ok I have it locally formatted exactly as it appears in the submission guidelines but it seems to be adding tabs that aren't there. Any suggestions?

![image](https://github.com/openjournals/joss-reviews/assets/72064522/4309b424-c2df-4ea9-affb-7205ec30a08a)
