### 2024-02-23 #6335 @b-pardi
@arfon thank you for suggesting the yaml linter, when I copied and pasted I realized the tabs are there, they just don't show up in my code editor, from some code formatter extension I have. I have disabled this and will be updating soon.
