### 2024-04-16 #6335 @b-pardi
> Diethelm Johannsmann: [johannsmann@pc.tu-clausthal.de](mailto:johannsmann@pc.tu-clausthal.de) ; DJohannsmann Julie Renner: [julie.renner@case.edu](mailto:julie.renner@case.edu) ; I am waiting on my PI to inform my of Julie's github, I will get back to you on that asap

@zhubonan  My PI has told me that Julie Renner was reached out to via email by someone from JOSS regarding being a reviewer, I am not sure on the status beyond that.
