### 2024-05-06 #6335 @zhubonan
Hi @DJohannsmann would you be happy to review this submission to JOSS? It appears to be very relevant to your field of work. Please also feel free to pass on this review invite to members of your team. Thanks in advance.

The review process basically involves reading the manuscript and more importantly checking and testing the software code. It has an open format - if you have any problems/questions/comments, you can interact with the author directly on GitHub under the review issue (which will be generated once the review process starts), or raise issues in the code's repository. If you need any assistance, please let me know. Thanks in advance.

The manuscript can be viewed here 👉 [View article proof on GitHub](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06335/joss.06335/10.21105.joss.06335.pdf)
Software repository 👉 https://github.com/b-pardi/BraTaDio
The JOSS reviewer policy and reviewer guideline for your reference 👉 [Reviewer Guideline](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
