### 2024-05-19 #6335 @b-pardi
Hi @zhubonan since we haven't gotten a response from others, I have chosen a few from the Joss reviewers list that mention having experience making data analysis tools with Python

https://github.com/simeonhebrew
https://github.com/jkanche
https://github.com/flekschas
