### 2024-05-19 #6335 @b-pardi
I also updated the paper, there was a slight issue with figure referencing
