### 2024-06-04 #6831 @editorialbot
[REVIEW]: pyQCM-BraTaDio: A tool for visualization, data mining, and modelling of Quartz crystal microbalance with dissipation data
**Submitting author:** @b-pardi (<a href="http://orcid.org/0000-0001-6483-9858">Brandon Pardi</a>)
**Repository:** https://github.com/b-pardi/BraTaDio
**Branch with paper.md** (empty if default branch):
**Version:** b.1.2.6
