### 2024-06-16 b-pardi/BraTaDio#1 @erik-whiting
JOSS Review Issues - first round
Hi @b-pardi, I'm Erik Whiting, one of the reviewers assigned by the Journal of Open Source Software, which is being manage on [this issue here](https://github.com/openjournals/joss-reviews/issues/6831). I am going off [the checklist generated by JOSS](https://github.com/openjournals/joss-reviews/issues/6831#issuecomment-2147616529). I've finished my initial review and found the following issues:

# Functionality

## Installation

There isn't much in the way of installation instructions. The README has _some_ information, but it's not very thorough and it seems more geared towards an edge case. Personally, I was able to figure out that installing and running `BraTaDio` meant cloning the repo and running `main.py`. I don't think you should assume general users will now this, you have to explicitly tell them how to run the software.

I will say, however, that after cloning and installing the requirements, the program ran perfectly.

## Functionality

I followed along with the instructions in the README and basically nothing happened. I tried a few things but my most recent run included the following steps:
1. Load the file `BraTaDio\sample_generations\qcmi-bsa-after\QSM-I-BSA_1mgpml`
2. Select `QCM-I`
3. Enter the number "10" for both t0 (s) and tf(s)
4. Select "theoretical"
5. Apply default values in plot options
6. Submit file information
7. See success message
8. Click both checkboxes: Raw Data Overtone Selection and Shifted Data Overtone Selection
9. Select "1st frequency" and "1st dissipation" in both
10. Clicked Submit
11. Noticed the following errors in my terminal (I'm not sure at which point they occurred though:
![image](https://github.com/b-pardi/BraTaDio/assets/47164199/d1a230ac-cfef-4253-964f-96916fe5cb32)
12. Nothing happened after clicking submit. Checked git for generated files, nothing there exception `__pycache__` generations (which you should probably include in your `.gitignore`.
13. Loaded the generated `Formatted-QSM-I-BSA_1mgpml` file
14. Nothing happened
15. Clicked "Submit file information"
16. Nothing happened
17. Left the same things that are currently "checked", Clicked "Submit"
18. Nothing happened

I may have done something wrong, or it's entirely possible I nothing was _**supposed**_ to happen. However, it's not clear from your instructions what I should expect to happen. Based on the README, I expected there to be a plot to look at somewhere, but I don't see it.

Also note that after filling everything out, I get the following errors in the console when trying to click any of the buttons in "modeling":
![image](https://github.com/b-pardi/BraTaDio/assets/47164199/fe56aa88-7d81-43c8-b50d-29c543ad31a9)

I can't really comment on anything else about the functionality since I cannot seem to make the graphs appear. Please advise on if I am misreading the instructions, if the instructions are unclear, or if there is a bug.

## Performance

I think this is N/A since you don't really make any performance claims.

# Documentation

## Statement of Need

You do cover this in the paper, but the documentation for the repository does not seem to address a statement of need for this software.

## Installation Instructions

As I mentioned above, the installation documentation is missing quite a bit.

## Automated Tests

These don't seem to be present at all. There are several files with "test" in the filename, but they are all in a directory called `deprecated`, so it is unclear if they are useful or not. There are also no CI files from what I can tell, so I don't think there is any automated tests.

This checkbox item in the reviewer guide also mentions manual tests for confirming functionality, I don't see any evidence of this anywhere either.

## Community Guidelines

There is a bug report file but there are no community guidelines or instructions for developers. I think the standard for these things is to include a `CONTRIBUTING` file to show people how to report issues and make pull requests, and a `.CODE_OF_CONDUCT` file for outlining how the community around your software is to behave.

***

Everything else looks pretty good, let me know if you have any questions. I will mark off the different items as they are addressed.
