### 2024-06-16 #6831 @erik-whiting
Hello @zhubonan, I have completed my initial review and opened an issue for my findings here: https://github.com/b-pardi/BraTaDio/issues/1
