### 2024-06-17 b-pardi/BraTaDio#1 @b-pardi
@erik-whiting thank you for taking the time out to review the software!
The first big thing I want to address is in your step 3, you say you put the same number for t0 and tf. This is an incorrect user since it is supposed to grab data between two different time to points to get the average baseline time. As mentioned in the File information section of the readme, for testing you can use an example range of 0-100 seconds. In practice, this is the last stable few minutes of the device finding an equilibrium before the experiment begins, and is determined during the experiment. This is an input I did not consider, and will act to add error catching to ensure there is a correctly entered time range.

Apologies for leaving out tests, this is my first user interactive project, so I did not know what to consider for tests, I will do some research and add those asap. Same goes for community guidelines, as the sole developer of the project, I was previously just zipping the project and sending updates to my team for use, and the report was made for non-developers to inform me of bugs. But with this in mind I will add more developer oriented reporting mechanisms. I also see your points on the shortcomings of the README and will address those as well.

I should have updates in the next few days, in the meantime please let me know how the rest of the software runs using the baseline time t0 and tf mentioned above.

Thank you again!
