### 2024-06-18 b-pardi/BraTaDio#1 @b-pardi
I believe I have taken care of all the documentation related issues (manual tests, installation instr, etc). I will need some time for the automated tests, but will update when I have that soon.
