### 2024-06-18 b-pardi/BraTaDio#1 @erik-whiting
I don't think purely automated tests are required, the checklist item states that documented manual tests to verify successful installation is good enough!
