### 2024-06-18 b-pardi/BraTaDio#2 @kstenio
Improve README and install instructions
Hi @b-pardi, sorry for taking this much time for a 1st round review. It took a lot of time for another reviewer to appear, and now I'm a little busier than before. Either way, here we are!

As another reviewer did with me while I was in your shoes, I am opting for opening multiple issues. So you can work on each one separately. This one is mostly regarding **install instructions**.

---

Your readme is fine, but could benefit from a finer structure. You can make a little introduction on what the software does and limitations. Something similar to the paper abstract.

Then, you should focus on a mini-tutorial on installing the application. The install_packages script seems totally unnecessary since I did not use it. I would recommend some code instructions or an ordered list of topics.

What I did:

```bash
# Cloned the repository
git clone https://github.com/b-pardi/BraTaDio/
# Entered the app folder
cd BraTaDio
# Created an Python virtual environment
pyton -m venv .venv
# Sourced the venv
source .venv/bin/activate
# Installed packages from requirements.txt
pip install -r requirements.txt
# Run the application
python main.py
```

Also worth mention that I am under **Linux** (Ubuntu 22.04 to be precise), and had no problems in setting up the environment, but I experienced a warning when generating graphics: `findfont: Font family 'Arial' not found`. This is a little obvious, as I do not have the **_Arial_** font on my system. I would recommend rely on font `family` other than `fontname` for [Matplotlib](https://matplotlib.org/stable/users/explain/text/text_props.html).

Finally, after the install instructions, you can resume your readme as it is now.
