### 2024-06-18 b-pardi/BraTaDio#3 @kstenio
Better description of authorship
In the paper, it is mentioned several (8) authors (including @b-pardi), but when it comes to the [commits](https://github.com/b-pardi/BraTaDio/commits/master/), you are the only contributor to the code.

Since the theme of the software is very specialized, I understand that maybe all those authors helped somehow in the software development, writing a piece of code or helping in the development in another manner. However I find very odd the statement that someone who did not uploaded any of the software's code to have contributed **equally** in its development (e.g. _Syeda Ahmed_).

In order to avoid misinterpretation, I recommend adding a proper **"Author contributions"** section before the _Acknowledgments_. You can follow the [CRediT (Contributor Roles Taxonomy)](https://www.elsevier.com/researcher/author/policies-and-guidelines/credit-author-statement) guidelines for support.

---

If it helps, I did a while ago a very simple script that reads a `xlsx` file and creates the **Author contributions** paragraph:

```python
import pandas as pd

def main(args):
	df = pd.read_excel('credit_matrix.xlsx', index_col=0, dtype=bool)
	paragraph = ''
	for cred in df.index:
		row = ''
		for auth in df.columns:
			if df.loc[cred, auth]:
				row += f' {auth},'
		paragraph += f'{cred}:{row[:-1]}; '
	with open('credit.txt', 'w') as cr:
		cr.write(paragraph[:-2]+'.')
	print(paragraph[:-2]+'.')
	return 0

if __name__ == '__main__':
	import sys
	sys.exit(main(sys.argv))
```

Where the `credit_matrix.xlsx` file is structured in the following manner:

| CRediT                         | A. A. | A. B. | A. C. | A. D. |
|--------------------------------|-------|-------|-------|-------|
| Conceptualization              | 1     | 1     | 0     | 1     |
| data curation                  | 1     | 1     | 0     | 0     |
| formal analysis                | 1     | 1     | 1     | 1     |
| funding acquisition            | 0     | 0     | 0     | 1     |
| investigation                  | 1     | 1     | 0     | 0     |
| methodology                    | 1     | 1     | 1     | 1     |
| project administration         | 1     | 0     | 0     | 1     |
| resources                      | 0     | 0     | 0     | 1     |
| software                       | 1     | 1     | 0     | 0     |
| supervision                    | 0     | 0     | 0     | 1     |
| validation                     | 1     | 1     | 1     | 1     |
| visualization                  | 1     | 1     | 1     | 0     |
| writing – original draft       | 1     | 1     | 0     | 0     |
| writing – review &amp; editing | 1     | 1     | 1     | 1     |

So, you can put in columns the initials of authors (A. A. == Author A, etc.), and a boolean for the contribution. Run the script, and it will create a `credit.txt` file with the proper data for you. :+1:
