### 2024-06-18 b-pardi/BraTaDio#4 @kstenio
Cite libraries papers when possible
You use NumPy, SciPy, Matplotlib and Pandas. All of these libraries have published papers/books with guidelines for citation. I recommend including then in the **References**.

As an example, [for the NumPy](https://numpy.org/citing-numpy/), you can use this citation:

```bib
@article{harris2020array,
  title={Array programming with NumPy},
  author={Harris, Charles R and Millman, K Jarrod and Van Der Walt, St{\'e}fan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
  journal={Nature},
  volume={585},
  number={7825},
  pages={357--362},
  year={2020},
  publisher={Nature Publishing Group UK London},
  doi = {10.1038/s41586-020-2649-2},
  url={https://doi.org/10.1038/s41586-020-2649-2}
}
```
