### 2024-06-18 b-pardi/BraTaDio#5 @kstenio
GUI could benefit from minor improvements
The GUI application is written in Tk, which is fine (although I would recommend checking out a more modern framework, like [PySide6](https://pypi.org/project/PySide6/)), but the size of the window is constant. If the user resolution is set to fullhd (1920x1080, a common standard nowadays), when maximizing the application, there is no issues. The same can not be said for smaller screens (e.g. 14" laptops). My recommendation is implementing properly responsivity. Non mandatory, but recommended. For what I've searched, [this](https://stackoverflow.com/questions/46522051/python-tkinter-responsive-grid) could be a start.

Another problem I've seen is regarding the input boxes. You used the same text input for all fields, but some are strictly numbers (baseline, visualize overtone, maybe others). For those, try using a [spinbox](https://www.tutorialspoint.com/python/tk_spinbox.htm).

Also, it is very unclear to me when the plots are computed. It took a while to understand that the plots are performed when clicking "submit". To me, "save graphics" or "perform plots" would be better. You could also show a messagebox, maybe with a link for the folder (html), when the process is finished. If you use `pathlib` library, it would be something like `folder_str = f'<a href={folder_path.as_uri()}>{folder_path.name}</a>'`

Finally, I noticed that it is not possible to deselect "interactive plot" radio button in raw and shifted data overtone selection.
