### 2024-06-18 b-pardi/BraTaDio#6 @kstenio
No (automated) tests
I understand that tests are commonly left behind when designing a new software (TDD is not for everyone, myself included). But they are necessary to guarantee proper coverage (e.g. unit testing), and are a big help when updating the software libraries, or adding a new feature.

For example, a new version of Pandas arrives and maybe this could break the software (happened with me once), or you want to update from Python 3.10 to 3.12. You just have to create a new environment and run the tests to know if everything is OK. If a test breaks, it means something needs a fix.

You could use the built-in unittest library, but I would recommend using [pytest](https://pypi.org/project/pytest/) instead, as it is simpler to use. For me, if you create at least these 3 tests, it is fine:

1. See if GUI is properly spawned
2. If the program loads sample data properly
3. If the program saves plots properly
