### 2024-06-18 b-pardi/BraTaDio#7 @kstenio
File structure could be cleaner (optional)
Your repository has multiple modules in the root folder. It is OK, but the most common is having a source code folder (usually named `src`) and a caller from the outer scope (root).

This is a **_suggestion_** (I created a `doc` and `src` folder, and also updated the imports to include the **src**, e.g. `from src.analyze import analyze_data, ordinal`)

```bash
.
├── deprecated
│   ├── dissipation_test.py
│   ├── fn_main.py
│   ├── main_old.py
│   ├── plotly_ex.py
│   ├── qcmi_dissipation_test.py
│   ├── QCMI-FULL-VALUES.png
│   ├── QCMi.png
│   ├── Qsense.png
│   ├── span_mems.txt
│   ├── span.py
│   ├── span_test2.py
│   ├── span_test.py
│   ├── tempgui.py
│   ├── test.png
│   ├── test.py
│   ├── text.out
│   └── voinova.py
├── doc
│   ├── basic_template.tex
│   ├── BraTaDio_example_bug_report.docx
│   └── pyQCM-BraTaDio SI.pdf
├── figs
│   ├── figure1.png
│   ├── figure2.png
│   ├── figure3.png
│   └── figure4.png
├── offset_data
│   ├── COPY-PASTE_OFFSET_VALUES_HERE.csv
│   ├── theoretical_frequencies.csv
│   └── theoretical_frequencies.xlsx
├── plot_opts
│   ├── default_opts.json
│   └── plot_customizations.json
├── qcmd-plots
│   ├── disp_V_freq_plot.png
│   ├── dissipation_plot.png
│   ├── freq_dis_V_time.png
│   ├── frequency_plot.png
│   ├── RAW-dissipation-plot.png
│   ├── RAW-resonant-freq-plot.png
│   └── temp_vs_time_plot.png
├── raw_data
│   └── Formatted-QSM-I-BSA_1mgpml.csv
├── res
│   └── m3b_comp.png
├── sample_generations
│   ├── qcmi-bsa-after
│   │   ├── crystal_thickness.png
│   │   ├── dissipation_plot.png
│   │   ├── frequency_plot.png
│   │   ├── QSM-I-BSA_1mgpml.csv
│   │   ├── RAW-dissipation-plot.png
│   │   ├── RAW-resonant-freq-plot.png
│   │   ├── Sauerbrey_fit_range_test.png
│   │   └── thin_film_liquid_test.png
│   └── qsense-bsa-after
│       ├── BSA.1mgml-1.280723_QSD.qsd
│       ├── BSA.1mgml-1.280723-unprotected.xlsx
│       ├── crystal_thickness.png
│       ├── dissipation_plot.png
│       ├── Formatted-BSA.1mgml-1.280723-unprotected.csv
│       ├── frequency_plot.png
│       ├── offset values.txt
│       ├── RAW-dissipation-plot.png
│       ├── RAW-resonant-freq-plot.png
│       ├── Sauerbrey_fit_range_test.png
│       └── thin_film_liquid_test.png
├── selected_ranges
│   ├── clean_all_stats_dis.csv
│   ├── clean_all_stats_rf.csv
│   ├── crystal_thickness_output.csv
│   ├── raw_all_stats_dis.csv
│   ├── raw_all_stats_rf.csv
│   ├── sauerbrey_output.csv
│   ├── Sauerbrey_ranges.csv
│   ├── Sauerbrey_stats.csv
│   ├── thin_film_air_output.csv
│   └── thin_film_liquid_output.csv
├── src
│   ├── analyze.py
│   ├── Exceptions.py
│   ├── format_file.py
│   ├── format_qsd.py
│   ├── __init__.py
│   └── modeling.py
├── web_pkg
│   ├── templates
│   │   └── upload.html
│   └── web_main.py
├── bratadio.py
├── CHANGELOG.md
├── LICENSE
├── paper.bib
├── paper.md
├── README.md
└── requirements.txt

15 directories, 82 files
```

Additionally, the tests could be added to a `src/tests` folder.

Another tip is using a template for the [.gitignore](https://github.com/github/gitignore/blob/main/Python.gitignore) file and add yours as an extra.
