### 2024-06-18 #6831 @kstenio
Hello @zhubonan , I also made my initial review.

---

#### Open Issues

* [Improve README and install instructions b-pardi/BraTaDio#2](https://github.com/b-pardi/BraTaDio/issues/2)
* [Better description of authorship b-pardi/BraTaDio#3](https://github.com/b-pardi/BraTaDio/issues/3)
* [Cite libraries papers when possible b-pardi/BraTaDio#4](https://github.com/b-pardi/BraTaDio/issues/4)
* [GUI could benefit from minor improvements b-pardi/BraTaDio#5](https://github.com/b-pardi/BraTaDio/issues/5)
* [No tests b-pardi/BraTaDio#6](https://github.com/b-pardi/BraTaDio/issues/6)
* [File structure could be cleaner (optional) b-pardi/BraTaDio#7](https://github.com/b-pardi/BraTaDio/issues/7)
