### 2024-06-18 b-pardi/BraTaDio#2 @b-pardi
Thank you @kstenio for getting to review it! No worries about taking some time to get to it, I appreciate it regardless. I have added a few more details like specifying how to setup a venv, but I also have responded to much of the feedback of the other reviewer as well, so let me know if the changes are satisfactory. I'll close the issue for now, but please let me know if there's more I need to do!
Also I do appreciate the multiple separate issues, it makes these much more manageable.
