### 2024-06-18 b-pardi/BraTaDio#3 @b-pardi
That's super helpful thank you!

There is only one piece of the project that was written by someone else, JM Friedt. He wrote the raw QSD file formatting bit manipulation code, everything else was me. And he emailed it to me unprompted which is why he hasn't actually committed anything lol, but with his permission and an offer of co authorship he was happy to have his code in the project.

As for Syeda Ahmed, you have a good point. Her contribution was data curation, and it was my PI's idea to give her equal contrib for that reason, but that is probably something I should discuss with him. I will have this updated as soon as I can meet with my PI and discuss contributions, thanks!
