### 2024-06-18 b-pardi/BraTaDio#6 @b-pardi
Thank you for that, this is my first big project so tests weren't something even in my mind, but you're right they are important, and will add those asap
