### 2024-06-18 b-pardi/BraTaDio#5 @b-pardi
Thank you for these improvements. I will focus on the more important required things for now, but these are things I will definitely implement shortly after.
