### 2024-06-18 b-pardi/BraTaDio#3 @kstenio
Authorship and acknowledgement are very important in papers. Everyone involved wants proper credit, and to me the Credit guidelines are a great help!

You are right in talking with your PI before any commits, no problem.

Also, by convention, the PI name is the last.
