### 2024-06-18 b-pardi/BraTaDio#2 @kstenio
I've checked the changes, and for me they are sufficient :-)
