### 2024-06-18 b-pardi/BraTaDio#6 @kstenio
> Thank you for that, this is my first big project so tests weren't something even in my mind, but you're right they are important, and will add those asap

Great. Do it on your terms. I recommend watching a pytest tutorial and, of course, the official documentation.

Also worth the check if there are any pytest plugins for Tk.
