### 2024-06-18 b-pardi/BraTaDio#7 @b-pardi
Thank you for the suggestions, I've been meaning to clean this up for a while but now is a good time.
