### 2024-06-18 b-pardi/BraTaDio#7 @kstenio
> Thank you for the suggestions, I've been meaning to clean this up for a while but now is a good time.

Wow, that was quick!

Did you delete the pyc files in the repository? (I am on GitHub app right now, a little hard to check)

Those files could raise an OSError when running pytest.
