### 2024-06-18 b-pardi/BraTaDio#1 @b-pardi
Well I'm glad to hear that! I still plan on adding at least a few to be as thorough as possible, but I'm glad they are not required to continue
