### 2024-06-19 b-pardi/BraTaDio#6 @kstenio
> So far I have found the best way is using the unittest patch to simulate button clicks, checkbox selections, etc. still using pytest for everything else, and I can just check if there are instances of certain things, and check if certain labels exist and if they say the right thing

That's a nice approach. And, so far, your 1st test seems to be on track.
