### 2024-06-20 b-pardi/BraTaDio#6 @b-pardi
tests should be operational and added to readme for instr on how to run them
from root Bratadio dir: `pytest src/test/`
pillow will need to be installed
