### 2024-06-20 b-pardi/BraTaDio#1 @b-pardi
I believe all things mentioned in this thread have been remedied, please open a new issue if there is anything more. Apologies for not mentioning issues in commits, I forgot that was thing until recently.
