### 2024-06-21 b-pardi/BraTaDio#5 @b-pardi
All UI adjustments should have been made. Please open another issue if something was missed.
