### 2024-06-21 b-pardi/BraTaDio#2 @b-pardi
@kstenio I closed this issue before I saw your last comment on the font issue. Can you tell me what line it was you got the issue? Cuz I believe I used fontfamily everywhere.
