### 2024-06-24 b-pardi/BraTaDio#2 @kstenio
It is a warning that appears when saving a graphic.

![image](https://github.com/b-pardi/BraTaDio/assets/9257942/d8643515-b5f0-460f-90da-7358dd4bead3)
`WARNING  matplotlib.font_manager:font_manager.py:1371 findfont: Font family 'Arial' not found.`
