### 2024-06-24 b-pardi/BraTaDio#3 @b-pardi
Met with my PI this morning and established authorship credits. Thank you again!
