### 2024-07-01 b-pardi/BraTaDio#3 @kstenio
Hello! @b-pardi , I was looking into the paper and did not found the `Author contributions` section, and then realized: you added the TXT file into the repository. Sorry not explaining properly. :sweat_smile:

The script above is just a helper; the section is supposed to be in the paper, prior to `Acknowledgements` section, like this one:

![image](https://github.com/b-pardi/BraTaDio/assets/9257942/e694d16c-e268-437d-b118-1c5ea8fc840c)

Just include the content of the `credit.txt` file into [paper.bib](https://github.com/b-pardi/BraTaDio/blob/master/paper.md), for example:

```markdown
(...) and crystal thickness. This tool is key for efficient data analysis in preference over laborious spreadsheet evaluation.

# Author contributions

Conceptualization: Brandon Pardi, Syeda Tajin Ahmed, Roberto C. Andresen Eguiluz; data curation: Syeda Tajin Ahmed, Silvia Jonguitud Flores, Warren Flores, Bernardo Yáñez Soto, Roberto C. Andresen Eguiluz; formal analysis: Brandon Pardi, Syeda Tajin Ahmed, Jean-Michel Friedt, Laura L.E. Mears, Bernardo Yáñez Soto, Roberto C. Andresen Eguiluz; funding acquisition: Bernardo Yáñez Soto, Roberto C. Andresen Eguiluz; investigation: Brandon Pardi, Syeda Tajin Ahmed, Roberto C. Andresen Eguiluz; methodology: Brandon Pardi, Syeda Tajin Ahmed, Jean-Michel Friedt, Bernardo Yáñez Soto, Roberto C. Andresen Eguiluz; project administration: Brandon Pardi, Roberto C. Andresen Eguiluz; resources: Brandon Pardi, Syeda Tajin Ahmed, Laura L.E. Mears, Bernardo Yáñez Soto, Roberto C. Andresen Eguiluz; software: Brandon Pardi, Jean-Michel Friedt; supervision: Brandon Pardi, Bernardo Yáñez Soto, Roberto C. Andresen Eguiluz; validation: Brandon Pardi, Syeda Tajin Ahmed, Silvia Jonguitud Flores, Warren Flores, Jean-Michel Friedt, Laura L.E. Mears, Bernardo Yáñez Soto, Roberto C. Andresen Eguiluz; visualization: Brandon Pardi, Syeda Tajin Ahmed, Laura L.E. Mears, Roberto C. Andresen Eguiluz; writing - original draft: Brandon Pardi, Syeda Tajin Ahmed, Silvia Jonguitud Flores, Bernardo Yáñez Soto, Roberto C. Andresen Eguiluz; writing - review & editing: Brandon Pardi, Syeda Tajin Ahmed, Silvia Jonguitud Flores, Laura L.E. Mears, Bernardo Yáñez Soto, Roberto C. Andresen Eguiluz.

# Acknowledgements

R.C.A.E. acknowledges funding from the NSF-CREST: Center for Cellular and Biomolecular Machines through the support of the National Science Foundation (NSF) Grant No. NSF-HRD-1547848. R.C.A.E. and B.P. acknowledge funding from the CAREER grant NSF CMMI Grant No. #2239665 awarded to R.C.A.E.

# Supporting information
The authors have compiled additional (...)
```

Finally, I have three extra recommendations (non mandatory):

1. Do you prefer the full name in `Author contributions`? I believe just initials are sufficient;
2. It is usual in papers adding a `Conflict of interest` section. The example above is fine;
3. The `---` separating the sections feels odd to me when reading the PDF, but it is up to you.
