### 2024-07-01 b-pardi/BraTaDio#2 @kstenio
Hello @b-pardi, any news on the above warning? I tried replacing "Arial" and "font" for "sans-serif" and "font-family" in some parts of the code (json files and analyze.py), but another issue happens. It would be nice fixing this, but since it is just a warning, I think it is OK. (but you should consider opening an issue later...)
