### 2024-07-01 b-pardi/BraTaDio#5 @kstenio
> All UI adjustments should have been made. Please open another issue if something was missed.

I've confirmed (at least for my configuration) that the program only quits if I click the "Exit" button inside application, but not if I click the "X" in the bar.

The program was opened from a terminal (`python main.py`) and I had to use Ctrl+Z to escape.
