### 2024-07-01 b-pardi/BraTaDio#6 @kstenio
Two tests are not working for me, either running `pytest` or `pytest src/test/`

```bash
platform linux -- Python 3.10.14, pytest-8.2.2, pluggy-1.5.0 -- /home/k/PycharmProjects/BraTaDio/.venv/bin/python
cachedir: .pytest_cache
rootdir: /home/k/PycharmProjects/BraTaDio
collecting ... collected 7 items

src/test/test_analyze.py::test_qcmi_plots FAILED                         [ 14%]
src/test/test_analyze.py::test_qsense_plots FAILED                       [ 28%]
src/test/test_format_file.py::test_basic_reading[sample_generations/qcmi-bsa-after/QSM-I-BSA_1mgpml.csv-format_QCMi] PASSED [ 42%]
src/test/test_format_file.py::test_basic_reading[sample_generations/qsense-bsa-after/BSA.1mgml-1.280723-unprotected.xlsx-format_Qsense] PASSED [ 57%]
src/test/test_format_file.py::test_qsense_file_formatting PASSED         [ 71%]
src/test/test_format_file.py::test_qsd_reading PASSED                    [ 85%]
src/test/test_ui.py::test_ui_spawns PASSED                               [100%]
```

Errors are:

1. **AssertionError**: assert False where False = plots_are_similar('sample_generations/qcmi-bsa-after/RAW-resonant-freq-plot.png', 'qcmd-plots/RAW-resonant-freq-plot.png')
2. **FileNotFoundError**: [Errno 2] No such file or directory: 'raw_data/Formatted-BSA.1mgml-1.280723-unprotected.csv'

And, this is the full log of running `pytest src/test/ -vvv`:

```bash
platform linux -- Python 3.10.14, pytest-8.2.2, pluggy-1.5.0 -- /home/k/PycharmProjects/BraTaDio/.venv/bin/python
cachedir: .pytest_cache
rootdir: /home/k/PycharmProjects/BraTaDio
collecting ... collected 7 items

src/test/test_analyze.py::test_qcmi_plots FAILED                         [ 14%]
src/test/test_analyze.py::test_qsense_plots FAILED                       [ 28%]
src/test/test_format_file.py::test_basic_reading[sample_generations/qcmi-bsa-after/QSM-I-BSA_1mgpml.csv-format_QCMi] PASSED [ 42%]
src/test/test_format_file.py::test_basic_reading[sample_generations/qsense-bsa-after/BSA.1mgml-1.280723-unprotected.xlsx-format_Qsense] PASSED [ 57%]
src/test/test_format_file.py::test_qsense_file_formatting PASSED         [ 71%]
src/test/test_format_file.py::test_qsd_reading PASSED                    [ 85%]
src/test/test_ui.py::test_ui_spawns PASSED                               [100%]

=================================== FAILURES ===================================
_______________________________ test_qcmi_plots ________________________________

init_qcmi_input_data = ('/home/k/PycharmProjects/BraTaDio/qcmd-plots/', <main.Input object at 0x7f2bf935b0a0>)

    def test_qcmi_plots(init_qcmi_input_data):
        # ensure plot dir clear before running
        plot_dir, qcmi_input = init_qcmi_input_data
        plot_dir = os.path.join(os.getcwd(), 'qcmd-plots/')

        analyze_data(qcmi_input)

        # check if plots generated
        dir_files = os.listdir(plot_dir)
        missing_files = [file for file in plot_filenames if file not in dir_files]
        assert len(missing_files) == 0, f"missing plots: {missing_files}"

        # check if plots match the sample generations
        for file in dir_files:
            if any(file.endswith(ext) for ext in img_exts):
                sample_file = os.path.join('sample_generations/qcmi-bsa-after/', file)
                generated_file = os.path.join('qcmd-plots/', file)
                print(sample_file,generated_file)
>               assert plots_are_similar(sample_file, generated_file)
E               AssertionError: assert False
E                +  where False = plots_are_similar('sample_generations/qcmi-bsa-after/RAW-resonant-freq-plot.png', 'qcmd-plots/RAW-resonant-freq-plot.png')

src/test/test_analyze.py:112: AssertionError
----------------------------- Captured stdout call -----------------------------
Using formatted file: raw_data/Formatted-QSM-I-BSA_1mgpml.csv
dataframe             Time  fundamental_freq  ...  13th_dis       Temp
0       5.054725       4960883.127  ...  0.000055  25.004446
1      10.113459       4960883.250  ...  0.000055  25.004490
2      15.164982       4960883.205  ...  0.000055  24.999670
3      20.212211       4960883.290  ...  0.000055  24.995724
4      25.267360       4960883.191  ...  0.000055  24.999746
..           ...               ...  ...       ...        ...
874  4415.497344       4960865.687  ...  0.000057  25.005959
875  4420.545194       4960865.663  ...  0.000056  24.991097
876  4425.592485       4960865.704  ...  0.000056  24.991081
877  4430.640259       4960865.669  ...  0.000056  25.009392
878  4435.681435       4960865.699  ...  0.000057  24.993872

[879 rows x 16 columns]
***

             Time  fundamental_freq  ...  13th_dis       Temp
0       5.054725       4960883.127  ...  0.000055  25.004446
1      10.113459       4960883.250  ...  0.000055  25.004490
2      15.164982       4960883.205  ...  0.000055  24.999670
3      20.212211       4960883.290  ...  0.000055  24.995724
4      25.267360       4960883.191  ...  0.000055  24.999746
..           ...               ...  ...       ...        ...
874  4415.497344       4960865.687  ...  0.000057  25.005959
875  4420.545194       4960865.663  ...  0.000056  24.991097
876  4425.592485       4960865.704  ...  0.000056  24.991081
877  4430.640259       4960865.669  ...  0.000056  25.009392
878  4435.681435       4960865.699  ...  0.000057  24.993872

[879 rows x 16 columns]

            Time       Temp
0       5.054725  25.004446
1      10.113459  25.004490
2      15.164982  24.999670
3      20.212211  24.995724
4      25.267360  24.999746
..           ...        ...
874  4415.497344  25.005959
875  4420.545194  24.991097
876  4425.592485  24.991081
877  4430.640259  25.009392
878  4435.681435  24.993872

[879 rows x 2 columns]

***
['fundamental_dis', '3rd_dis', '5th_dis', '7th_dis', '9th_dis', '11th_dis', '13th_dis'] ********
********             Time  fundamental_freq  fundamental_dis
0      10.113459       4960883.250         0.000187
1      15.164982       4960883.205         0.000187
2      20.212211       4960883.290         0.000187
3      25.267360       4960883.191         0.000187
4      30.314913       4960883.245         0.000187
..           ...               ...              ...
873  4415.497344       4960865.687         0.000187
874  4420.545194       4960865.663         0.000187
875  4425.592485       4960865.704         0.000187
876  4430.640259       4960865.669         0.000187
877  4435.681435       4960865.699         0.000187

[878 rows x 3 columns]
clean freq ch: fundamental_freq; clean disp ch: fundamental_dis
clean freq ch: fundamental_freq; clean disp ch: fundamental_dis
rf average: 4960883.211222221; dis average: 0.00018685483333333333

['fundamental_dis', '3rd_dis', '5th_dis', '7th_dis', '9th_dis', '11th_dis', '13th_dis'] ********
********             Time     3rd_freq   3rd_dis
0      10.113459  14866043.66  0.000099
1      15.164982  14866043.70  0.000099
2      20.212211  14866043.85  0.000099
3      25.267360  14866043.61  0.000099
4      30.314913  14866043.64  0.000099
..           ...          ...       ...
873  4415.497344  14865989.49  0.000100
874  4420.545194  14865989.62  0.000100
875  4425.592485  14865989.59  0.000100
876  4430.640259  14865989.64  0.000100
877  4435.681435  14865989.60  0.000100

[878 rows x 3 columns]
clean freq ch: 3rd_freq; clean disp ch: 3rd_dis
clean freq ch: 3rd_freq; clean disp ch: 3rd_dis
rf average: 14866043.578333333; dis average: 9.939444444444444e-05

['fundamental_dis', '3rd_dis', '5th_dis', '7th_dis', '9th_dis', '11th_dis', '13th_dis'] ********
********             Time     5th_freq   5th_dis
0      10.113459  24773492.65  0.000078
1      15.164982  24773492.73  0.000078
2      20.212211  24773493.03  0.000078
3      25.267360  24773492.93  0.000078
4      30.314913  24773493.12  0.000078
..           ...          ...       ...
873  4415.497344  24773405.84  0.000079
874  4420.545194  24773406.07  0.000079
875  4425.592485  24773406.38  0.000079
876  4430.640259  24773406.08  0.000079
877  4435.681435  24773406.16  0.000079

[878 rows x 3 columns]
clean freq ch: 5th_freq; clean disp ch: 5th_dis
clean freq ch: 5th_freq; clean disp ch: 5th_dis
rf average: 24773492.817222223; dis average: 7.85e-05

['fundamental_dis', '3rd_dis', '5th_dis', '7th_dis', '9th_dis', '11th_dis', '13th_dis'] ********
********             Time     7th_freq   7th_dis
0      10.113459  34679595.23  0.000067
1      15.164982  34679595.38  0.000067
2      20.212211  34679595.01  0.000067
3      25.267360  34679594.46  0.000067
4      30.314913  34679595.30  0.000067
..           ...          ...       ...
873  4415.497344  34679472.18  0.000068
874  4420.545194  34679472.83  0.000068
875  4425.592485  34679472.84  0.000068
876  4430.640259  34679472.74  0.000068
877  4435.681435  34679471.61  0.000068

[878 rows x 3 columns]
clean freq ch: 7th_freq; clean disp ch: 7th_dis
clean freq ch: 7th_freq; clean disp ch: 7th_dis
rf average: 34679594.78; dis average: 6.718333333333332e-05

['fundamental_dis', '3rd_dis', '5th_dis', '7th_dis', '9th_dis', '11th_dis', '13th_dis'] ********
********             Time     9th_freq   9th_dis
0      10.113459  44586432.24  0.000062
1      15.164982  44586431.69  0.000062
2      20.212211  44586430.13  0.000062
3      25.267360  44586432.62  0.000061
4      30.314913  44586434.81  0.000062
..           ...          ...       ...
873  4415.497344  44586273.32  0.000063
874  4420.545194  44586276.06  0.000063
875  4425.592485  44586275.08  0.000063
876  4430.640259  44586273.94  0.000063
877  4435.681435  44586274.93  0.000063

[878 rows x 3 columns]
clean freq ch: 9th_freq; clean disp ch: 9th_dis
clean freq ch: 9th_freq; clean disp ch: 9th_dis
rf average: 44586431.68611111; dis average: 6.157777777777778e-05

['fundamental_dis', '3rd_dis', '5th_dis', '7th_dis', '9th_dis', '11th_dis', '13th_dis'] ********
********             Time    11th_freq  11th_dis
0      10.113459  54493088.92  0.000057
1      15.164982  54493087.13  0.000057
2      20.212211  54493089.31  0.000057
3      25.267360  54493088.08  0.000057
4      30.314913  54493087.70  0.000057
..           ...          ...       ...
873  4415.497344  54492891.20  0.000059
874  4420.545194  54492891.35  0.000059
875  4425.592485  54492891.53  0.000059
876  4430.640259  54492890.71  0.000059
877  4435.681435  54492892.05  0.000059

[878 rows x 3 columns]
clean freq ch: 11th_freq; clean disp ch: 11th_dis
clean freq ch: 11th_freq; clean disp ch: 11th_dis
rf average: 54493087.528888896; dis average: 5.706111111111111e-05

['fundamental_dis', '3rd_dis', '5th_dis', '7th_dis', '9th_dis', '11th_dis', '13th_dis'] ********
********             Time    13th_freq  13th_dis
0      10.113459  64400341.31  0.000055
1      15.164982  64400341.21  0.000055
2      20.212211  64400342.83  0.000055
3      25.267360  64400342.56  0.000055
4      30.314913  64400339.32  0.000055
..           ...          ...       ...
873  4415.497344  64400103.89  0.000057
874  4420.545194  64400105.90  0.000056
875  4425.592485  64400108.68  0.000056
876  4430.640259  64400110.08  0.000056
877  4435.681435  64400110.17  0.000057

[878 rows x 3 columns]
clean freq ch: 13th_freq; clean disp ch: 13th_dis
clean freq ch: 13th_freq; clean disp ch: 13th_dis
rf average: 64400341.33944445; dis average: 5.473333333333334e-05

            Time  fundamental_freq  ...  13th_dis       Temp
0      10.113459       4960883.250  ...  0.000055  25.004490
1      15.164982       4960883.205  ...  0.000055  24.999670
2      20.212211       4960883.290  ...  0.000055  24.995724
3      25.267360       4960883.191  ...  0.000055  24.999746
4      30.314913       4960883.245  ...  0.000055  25.006128
..           ...               ...  ...       ...        ...
873  4415.497344       4960865.687  ...  0.000057  25.005959
874  4420.545194       4960865.663  ...  0.000056  24.991097
875  4425.592485       4960865.704  ...  0.000056  24.991081
876  4430.640259       4960865.669  ...  0.000056  25.009392
877  4435.681435       4960865.699  ...  0.000057  24.993872

[878 rows x 16 columns]
*** Plots Generated ***
sample_generations/qcmi-bsa-after/RAW-resonant-freq-plot.png qcmd-plots/RAW-resonant-freq-plot.png
0.06944710995148463
------------------------------ Captured log call -------------------------------
WARNING  matplotlib.font_manager:font_manager.py:1432 findfont: Font family ['Arial'] not found. Falling back to DejaVu Sans.
______________________________ test_qsense_plots _______________________________

init_qsense_input_data = ('/home/k/PycharmProjects/BraTaDio/qcmd-plots/', <main.Input object at 0x7f2bf634e2f0>)

    def test_qsense_plots(init_qsense_input_data):
        # ensure plot dir clear before running
        plot_dir, qsense_input = init_qsense_input_data
        plot_dir = os.path.join(os.getcwd(), 'qcmd-plots/')

>       analyze_data(qsense_input)

src/test/test_analyze.py:119:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
src/analyze.py:773: in analyze_data
    df = pd.read_csv(analysis.formatted_fn)
../../.pyenv/versions/3.10.14/lib/python3.10/site-packages/pandas/util/_decorators.py:311: in wrapper
    return func(*args, **kwargs)
../../.pyenv/versions/3.10.14/lib/python3.10/site-packages/pandas/io/parsers/readers.py:680: in read_csv
    return _read(filepath_or_buffer, kwds)
../../.pyenv/versions/3.10.14/lib/python3.10/site-packages/pandas/io/parsers/readers.py:575: in _read
    parser = TextFileReader(filepath_or_buffer, **kwds)
../../.pyenv/versions/3.10.14/lib/python3.10/site-packages/pandas/io/parsers/readers.py:934: in __init__
    self._engine = self._make_engine(f, self.engine)
../../.pyenv/versions/3.10.14/lib/python3.10/site-packages/pandas/io/parsers/readers.py:1218: in _make_engine
    self.handles = get_handle(  # type: ignore[call-overload]
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

path_or_buf = 'raw_data/Formatted-BSA.1mgml-1.280723-unprotected.csv'
mode = 'r'

    @doc(compression_options=_shared_docs["compression_options"] % "path_or_buf")
    def get_handle(
        path_or_buf: FilePath | BaseBuffer,
        mode: str,
        *,
        encoding: str | None = None,
        compression: CompressionOptions = None,
        memory_map: bool = False,
        is_text: bool = True,
        errors: str | None = None,
        storage_options: StorageOptions = None,
    ) -> IOHandles[str] | IOHandles[bytes]:
        """
        Get file handle for given path/buffer and mode.

        Parameters
        ----------
        path_or_buf : str or file handle
            File path or object.
        mode : str
            Mode to open path_or_buf with.
        encoding : str or None
            Encoding to use.
        {compression_options}

            .. versionchanged:: 1.0.0
               May now be a dict with key 'method' as compression mode
               and other keys as compression options if compression
               mode is 'zip'.

            .. versionchanged:: 1.1.0
               Passing compression options as keys in dict is now
               supported for compression modes 'gzip', 'bz2', 'zstd' and 'zip'.

            .. versionchanged:: 1.4.0 Zstandard support.

        memory_map : bool, default False
            See parsers._parser_params for more information.
        is_text : bool, default True
            Whether the type of the content passed to the file/buffer is string or
            bytes. This is not the same as `"b" not in mode`. If a string content is
            passed to a binary file/buffer, a wrapper is inserted.
        errors : str, default 'strict'
            Specifies how encoding and decoding errors are to be handled.
            See the errors argument for :func:`open` for a full list
            of options.
        storage_options: StorageOptions = None
            Passed to _get_filepath_or_buffer

        .. versionchanged:: 1.2.0

        Returns the dataclass IOHandles
        """
        # Windows does not default to utf-8. Set to utf-8 for a consistent behavior
        encoding = encoding or "utf-8"

        # read_csv does not know whether the buffer is opened in binary/text mode
        if _is_binary_mode(path_or_buf, mode) and "b" not in mode:
            mode += "b"

        # validate encoding and errors
        codecs.lookup(encoding)
        if isinstance(errors, str):
            codecs.lookup_error(errors)

        # open URLs
        ioargs = _get_filepath_or_buffer(
            path_or_buf,
            encoding=encoding,
            compression=compression,
            mode=mode,
            storage_options=storage_options,
        )

        handle = ioargs.filepath_or_buffer
        handles: list[BaseBuffer]

        # memory mapping needs to be the first step
        handle, memory_map, handles = _maybe_memory_map(
            handle,
            memory_map,
            ioargs.encoding,
            ioargs.mode,
            errors,
            ioargs.compression["method"] not in _compression_to_extension,
        )

        is_path = isinstance(handle, str)
        compression_args = dict(ioargs.compression)
        compression = compression_args.pop("method")

        # Only for write methods
        if "r" not in mode and is_path:
            check_parent_directory(str(handle))

        if compression:
            if compression != "zstd":
                # compression libraries do not like an explicit text-mode
                ioargs.mode = ioargs.mode.replace("t", "")
            elif compression == "zstd" and "b" not in ioargs.mode:
                # python-zstandard defaults to text mode, but we always expect
                # compression libraries to use binary mode.
                ioargs.mode += "b"

            # GZ Compression
            if compression == "gzip":
                if is_path:
                    assert isinstance(handle, str)
                    # error: Incompatible types in assignment (expression has type
                    # "GzipFile", variable has type "Union[str, BaseBuffer]")
                    handle = gzip.GzipFile(  # type: ignore[assignment]
                        filename=handle,
                        mode=ioargs.mode,
                        **compression_args,
                    )
                else:
                    handle = gzip.GzipFile(
                        # No overload variant of "GzipFile" matches argument types
                        # "Union[str, BaseBuffer]", "str", "Dict[str, Any]"
                        fileobj=handle,  # type: ignore[call-overload]
                        mode=ioargs.mode,
                        **compression_args,
                    )

            # BZ Compression
            elif compression == "bz2":
                # No overload variant of "BZ2File" matches argument types
                # "Union[str, BaseBuffer]", "str", "Dict[str, Any]"
                handle = bz2.BZ2File(  # type: ignore[call-overload]
                    handle,
                    mode=ioargs.mode,
                    **compression_args,
                )

            # ZIP Compression
            elif compression == "zip":
                # error: Argument 1 to "_BytesZipFile" has incompatible type "Union[str,
                # BaseBuffer]"; expected "Union[Union[str, PathLike[str]],
                # ReadBuffer[bytes], WriteBuffer[bytes]]"
                handle = _BytesZipFile(
                    handle, ioargs.mode, **compression_args  # type: ignore[arg-type]
                )
                if handle.mode == "r":
                    handles.append(handle)
                    zip_names = handle.namelist()
                    if len(zip_names) == 1:
                        handle = handle.open(zip_names.pop())
                    elif len(zip_names) == 0:
                        raise ValueError(f"Zero files found in ZIP file {path_or_buf}")
                    else:
                        raise ValueError(
[…]
