### 2024-07-02 b-pardi/BraTaDio#3 @b-pardi
@kstenio  you know it makes a lot more sense now that you mention that 😆 Should be taken care of.
For the other issues I think I prefer contributions be full names personally, so I don't need to keep scrolling up to the author list in the beginning and figure out who has what initials. The rest of the points you made are solid and I corrected those as well
