### 2024-07-02 b-pardi/BraTaDio#2 @b-pardi
@kstenio I went ahead and opened a new issue [#8](https://github.com/b-pardi/BraTaDio/issues/8)
