### 2024-07-02 b-pardi/BraTaDio#6 @b-pardi
@kstenio When I made these test scripts I had the formatted data files from previous testing in the correct directory already, but the test code was not putting them where they needed to be. Should be remedied now
