### 2024-07-03 b-pardi/BraTaDio#3 @kstenio
Nice! The latest [paper artifact](https://github.com/b-pardi/BraTaDio/actions/runs/9769449367/artifacts/1661781095) is great! :smiley:
