### 2024-07-03 b-pardi/BraTaDio#6 @kstenio
Two tests are still breaking, but this time one of them shows a different error:
![image](https://github.com/b-pardi/BraTaDio/assets/9257942/48c8d4cd-6dc0-4f0f-8afc-eba0fcf11eee)

It looks like an issue while copying or cleaning folders, or maybe your environment have local static files, therefore your test are passing...
