### 2024-07-03 b-pardi/BraTaDio#6 @b-pardi
I think I may know the issue, @kstenio  can you check in the plot_opts/plot_customizations.json and tell me what yours says for fig_dpi? I believe I corrected all data clean up so that data files and figures are removed before and after plotting, but I think the figure DPI may be different on yours. I will add to the test code to control the DPI and not just reference from plot_customizations.json, but in the meantime please change fig_dpi to 200 if it is not already, as that is the dpi the sample figures are generated with.

EDIT: Now that I think about it, any non default setting could cause an issue. I'll see about setting the plot options to default settings before testing, and changing defaults to be more cross OS compatible
