### 2024-07-03 b-pardi/BraTaDio#6 @b-pardi
I believe I have made the appropriate adjustments to plot settings.
