### 2024-07-05 b-pardi/BraTaDio#6 @kstenio
Great! Tests are working now!

![image](https://github.com/b-pardi/BraTaDio/assets/9257942/426b1e50-da9e-4219-aa84-987213023f8f)

Now I'm moving forward to inform the editor that my review is finished. Great job! :smiley:
