### 2024-07-05 #6831 @zhubonan
Hi @b-pardi  there are a few more steps before I can move the submission to the next stage. Please make a release of the package and deposit to a publicly available repository (for example, Zenodo) to get an DOI. This DOI will be linked to the JOSS publication. Please see the checklist above and make sure the metadata matches with that of the paper.

In the manuscript text, there are a few things to be corrected/standardised. I have found the following issues:

- The spaces between text and citations are missing. Please always use a space in between.
- The maths symbols in the figure captions are not italic, as they should be.
- There is a missing `.` at line 41 (paper draft) on the first page between `two packages` and `Here`.
- Some figures are not cross-referenced, such as those at line 74, 75 and 80. Please check all cross references.
- `MatPlotLib` should be `Matplotlib at line 99.

Please also give the manuscript another round of proof and check for other errors. Thanks.

PS: Other recently published JOSS papers can act as examples on how the archive DOI is linked with the paper.
