### 2024-07-05 #6831 @b-pardi
@zhubonan I have a draft of the [Zenodo archive](https://zenodo.org/uploads/12668677?token=eyJhbGciOiJIUzUxMiIsImlhdCI6MTcyMDIyMTMyNCwiZXhwIjoxNzIwNzQyMzk5fQ.eyJpZCI6Ijg0NTU3MjFmLWVmNmUtNDZkYS04Mzc4LTUyY2U1N2M0ZWYwMSIsImRhdGEiOnt9LCJyYW5kb20iOiI5NmM1OTllYzAzZjFiNjY2ZThhN2YzMjViYmQ1ZTgyMiJ9.Uk-arpNC7N55soP12BMDrF_PYWcIFqYfXbILJ_PGiaCQ2RqceLKmaGHJuGs7S6tzsLu8vGVDBox_G23SS7E8Xg) if you would like to check it out and verify I filled out the required entries. I wasn't sure what to put for the Journal section as it hasn't been published yet. I assume I can go back and update that once it has been published?
