### 2024-07-10 b-pardi/BraTaDio#6 @kstenio
> Thank you! You've been a big help with this process.

Thanks for the kind words! But, I just did for you the same another reviewer from JOSS did for me a while ago. The whole point of this process is increasing the overall quality of paper **_and_** software.

You are now in the final step, takes a while, but worth it. :smiley:
