### 2024-07-12 #6831 @zhubonan
@b-pardi  Seems you have a few DOIs missing. Could you please update the bib file to include those?
