### 2024-07-12 #6831 @b-pardi
> The affiliations of the Zenodo archive appear to be different from that of the paper. Please check and update if necessary. You should be able to update the metadata of the Zenodo archive without making a new release (new DOI).

@zhubonan  Do they need to be matching word for word? When I setup the archive I would start typing in their affiliations and choose the one that came up. They are all the same place just worded differently.

I will fix the DOI issue now
