### 2024-07-14 #6831 @b-pardi
> > @zhubonan Do they need to be matching word for word? When I setup the archive I would start typing in their affiliations and choose the one that came up. They are all the same place just worded differently.
>
> Ideally yes to avoid any confusion as the paper is linked to the archive.

@zhubonan ok, should be fixed
