### 2024-07-17 #6831 @Kevin-Mattheus-Moerman
@b-pardi next I step in and will help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, however the below are some points that require your attention:

- [ ] On your GitHub repository I do not see a tagged release. Please create one. The label for this release should match what we have set here, i.e. `b.1.2.6`. If you alter the tag label then we need to edit it here and also on ZENODO as all need to match.
